Modify: model 명 및 import
model 명을 일부 수정하고 불필요한 improt 정리
