+++ src/main/java/com/snowball/boardadmin/domain/post/model/Post.java
@@ -1,9 +1,8 @@
-package com.snowball.boardadmin.domain.board.model;
+package com.snowball.boardadmin.domain.post.model;
 
+import com.snowball.boardadmin.common.util.UserRole;
 import lombok.*;
-
 import java.sql.Timestamp;
-
 /**
  * author         : Jason Lee
  * date           : 2023-07-13
@@ -15,12 +14,13 @@ import java.sql.Timestamp;
 @ToString
 @EqualsAndHashCode
 @Builder
-public class Comment {
+public class Post {
 
     private Long id;
     private Long userId;
-    private Long postId;
-    private String content;
+    private UserRole title;
+    private UserRole content;
+    private Boolean blindState;
     private Timestamp createdAt;
     private Timestamp updatedAt;
 }
