+++ src/main/java/com/snowball/boardadmin/domain/user/model/User.java
@@ -19,11 +19,11 @@ public class User {
 
     private Long id;
     // 유저의 Unique ID (로그인 ID)
-    private String userAccount;
-    private String userName;
-    private String password;
-    private String email;
-    private String nickName;
+    private java.lang.String userAccount;
+    private java.lang.String userName;
+    private java.lang.String password;
+    private java.lang.String email;
+    private java.lang.String nickName;
     // TRUE(Activated), FALSE(UnActivated)
     private Boolean userStatus;
     // 회원의 등급
